requirements.txt
django
python-dotenv
djangorestframework
djangorestframework-simplejwt